+++ src/main/java/org/opentripplanner/standalone/config/updaters/sources/BikeRentalSourceFactory.java
@@ -58,10 +58,29 @@ public class BikeRentalSourceFactory {
 
   public BikeRentalDataSourceParameters create() {
     switch (type) {
-      case GBFS: return new GbfsBikeRentalDataSourceParameters(url(), network(), routeAsCar(), allowKeepingBicycleRentalsAtDestination());
-      case KML:  return new GenericKmlBikeRentalDataSourceParameters(url(), namePrefix());
-      default:   return new BikeRentalDataSourceParameters(type, url(), network(), apiKey());
+      case GBFS:
+        return new GbfsBikeRentalDataSourceParameters(
+            url(),
+            network(),
+            routeAsCar(),
+            allowKeepingBicycleRentalsAtDestination(),
+            headers()
+        );
+      case KML:
+        return new GenericKmlBikeRentalDataSourceParameters(url(), namePrefix());
+      default:
+        return new BikeRentalDataSourceParameters(
+            type,
+            url(),
+            network(),
+            apiKey(),
+            headers()
+        );
     }
+  }
+
+  private Map<String, String> headers() {
+    return c.asMap("headers", NodeAdapter::asText);
   }
 
   private String url() {