+++ src/test/java/org/opentripplanner/standalone/config/NodeAdapterTest.java
@@ -310,4 +310,17 @@ public class NodeAdapterTest {
         );
         assertNull(subject.asLinearFunction("no-key", null));
     }
+
+    @Test
+    public void asMap() {
+        NodeAdapter subject = newNodeAdapterForTest("{ key : { A: true, B: false } }");
+        assertEquals(
+            Map.of("A", true, "B", false),
+            subject.asMap("key", NodeAdapter::asBoolean)
+        );
+        assertEquals(
+            Collections.<String, Boolean>emptyMap(),
+            subject.asMap("missing-key", NodeAdapter::asBoolean)
+        );
+    }
 }